rename file to getParentMap
